Move widget files to widgets folder
All nested under one folder.
