+++ src/components/grid/index.tsx
@@ -4,9 +4,9 @@ import { useScramble } from "use-scramble";
 import { WidgetGridContext } from "./hooks";
 import { GridProps } from "./props";
 import clsx from "clsx";
-import Twitter from "@/components/widgets/twitter";
-import Media from "@/components/widgets/media";
-import Project from "@/components/widgets/project";
+import Twitter from "@/components/widgets/variants/twitter";
+import Media from "@/components/widgets/variants/media";
+import Project from "@/components/widgets/variants/project";
 import type { Home, Clients } from "@/sanity/types";
 import { useEffect, useRef } from "react";
 import { useViewport } from "@/hooks/useViewport";
