+++ src/components/widgets/grid/base-widget.tsx
